+++ codigo/dispositivos/pantalla.c++
@@ -1,14 +1,14 @@
 #include "pantalla.h++"
 
-int Pantalla::Ancho = 0;
-int Pantalla::Alto = 0;
+float Pantalla::Ancho = 0;
+float Pantalla::Alto = 0;
 
-int Pantalla::Centro_horizontal()
+float Pantalla::Centro_horizontal()
 {
 	return Ancho / 2;
 }
 
-int Pantalla::Centro_vertical()
+float Pantalla::Centro_vertical()
 {
 	return Alto / 2;
 }
