+++ codigo/ventanas/ventana_configuracion.c++
@@ -36,6 +36,6 @@ void VentanaConfiguracion::evento_teclado(Tecla tecla, bool estado)
 		m_accion = CambiarATitulo;
 }
 
-void VentanaConfiguracion::evento_pantalla(int /*ancho*/, int /*alto*/)
+void VentanaConfiguracion::evento_pantalla(float /*ancho*/, float /*alto*/)
 {
 }
